+++ core/src/builtin_widgets/padding.rs
@@ -26,24 +26,23 @@ impl Render for Padding {
     // Shrink the clamp of child.
     let child_clamp = BoxClamp { min, max };
     ctx.force_child_relayout(child);
-    let mut child_layouter = ctx.assert_single_child_layouter();
+    let child = ctx.assert_single_child();
 
-    let mut size = child_layouter.perform_widget_layout(child_clamp);
-    if child_layouter.has_child() {
+    let mut size = ctx.perform_child_layout(child, child_clamp);
+    if child.first_child(ctx.tree).is_some() {
       // Expand the size, so the child have padding.
       size = clamp.clamp(size + thickness);
-      child_layouter.update_size(child, size);
+      ctx.update_size(child, size);
 
       // Update child's children position, let they have a correct position after
       // expanded with padding. padding.
-      let mut grandson_layouter = child_layouter.into_first_child_layouter();
-      while let Some(mut l) = grandson_layouter {
-        if let Some(pos) = l.box_pos() {
+      let mut ctx = LayoutCtx { id: child, tree: ctx.tree };
+      let (ctx, grandson) = ctx.split_children();
+      for g in grandson {
+        if let Some(pos) = ctx.widget_box_pos(g) {
           let pos = pos + Vector::new(self.padding.left, self.padding.top);
-          l.update_position(pos);
+          ctx.update_position(g, pos);
         }
-
-        grandson_layouter = l.into_next_sibling()
       }
     }