+++ core/src/widget_tree.rs
@@ -86,8 +86,8 @@ impl WidgetTree {
           .map(|info| info.clamp)
           .unwrap_or_else(|| BoxClamp { min: Size::zero(), max: win_size });
 
-        let mut layouter = Layouter::new(wid, true, self);
-        layouter.perform_widget_layout(clamp);
+        let mut ctx = LayoutCtx { id: wid, tree: self };
+        ctx.perform_child_layout(wid, clamp);
       }
     }
   }
@@ -242,12 +242,12 @@ pub(crate) struct Root;
 impl Render for Root {
   fn perform_layout(&self, clamp: BoxClamp, ctx: &mut LayoutCtx) -> Size {
     let mut size = ZERO_SIZE;
-    let mut layouter = ctx.first_child_layouter();
-    while let Some(mut l) = layouter {
-      let child_size = l.perform_widget_layout(clamp);
+    let (ctx, children) = ctx.split_children();
+    for c in children {
+      let child_size = ctx.perform_child_layout(c, clamp);
       size = size.max(child_size);
-      layouter = l.into_next_sibling();
     }
+
     size
   }